% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphantasiaCognitiveClustering-package.R
\docType{package}
\name{aphantasiaCognitiveClustering-package}
\alias{aphantasiaCognitiveClustering}
\alias{aphantasiaCognitiveClustering-package}
\title{aphantasiaCognitiveClustering: Data Analysis for "Unsupervised clustering reveals spatial and verbal cognitive profiles in aphantasia and typical imagery"}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains all the code to reproduce the analyses reported in the article "Unsupervised clustering reveals spatial and verbal cognitive profiles in aphantasia and typical imagery".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-delem.github.io/aphantasiaCognitiveClustering/}
  \item \url{https://github.com/m-delem/aphantasiaCognitiveClustering}
}

}
\author{
\strong{Maintainer}: Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

}
\keyword{internal}
