% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_selected_vars.R
\name{cluster_selected_vars}
\alias{cluster_selected_vars}
\title{Use Gaussian Mixture Models to cluster selected variables}
\usage{
cluster_selected_vars(df, vars = NULL)
}
\arguments{
\item{df}{A data frame containing the variables to be clustered, usually
\code{study_data}.}

\item{vars}{A character vector of variable names to be clustered. If NULL, a
default set of variables is used:
\itemize{
\item \code{visual_imagery}
\item \code{sensory_imagery}
\item \code{spatial_imagery}
\item \code{verbal_strategies}
\item \code{fluid_intelligence}
\item \code{verbal_reasoning}
\item \code{span_spatial}
}}
}
\value{
A fitted Mclust object containing the clustering results.
}
\description{
Use Gaussian Mixture Models to cluster selected variables
}
\examples{
clustering <- cluster_selected_vars(study_data)
clustering

if (requireNamespace("dplyr")) {
  data.frame(cluster = clustering$classification) |>
    dplyr::reframe(n = dplyr::n(), .by = cluster)
}
}
