% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_vars.R
\name{correlate_vars}
\alias{correlate_vars}
\title{Correlate the original variables with a chosen method and correction}
\usage{
correlate_vars(
  df,
  method = "pearson",
  partial = TRUE,
  correction = "bonferroni"
)
}
\arguments{
\item{df}{A data frame containing the variables to be correlated, usually
\code{study_data}.}

\item{method}{A string indicating the correlation method to use. Default is
"pearson". Other options include "spearman" and "kendall". See
\code{?correlation::correlation()} for more details.}

\item{partial}{Logical. Whether to compute partial correlations. Default is
TRUE. If FALSE, computes simple correlations.}

\item{correction}{A string indicating the p-value correction method to use.
Default is "bonferroni". Other options include "holm", "fdr", or "none".
See \code{?correlation::correlation()} for more details.}
}
\value{
A data frame containing the correlations between the variables, with
the parameters renamed for better readability.
}
\description{
Correlate the original variables with a chosen method and correction
}
\examples{
simple_correlations <- correlate_vars(study_data, partial = FALSE)
format(head(simple_correlations))
}
