% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_study_variables.R
\name{filter_study_variables}
\alias{filter_study_variables}
\title{Filter the \code{Variable} column of a  long format data frame}
\usage{
filter_study_variables(df, selection = "original", variables = NULL)
}
\arguments{
\item{df}{A data frame in long format containing the \code{Variable} column to be
filtered. This is for example the output of \code{get_longer(study_data)}.}

\item{selection}{A character string indicating the selection of variables to
filter for. It can be either \code{"original"}, \code{"reduced"} (reduced variables +
three validation variables from the original set) or \code{"reduced_strict"}.
If you want to use a custom set of variables, set \code{variables} to a character
vector and leave \code{selection} to its default value.}

\item{variables}{Optional. A character vector of variable names to filter
for. If \code{NULL}, the function will use the default selection of variables
based on the \code{selection} argument. If provided, it will override the
default selection.}
}
\value{
A data frame filtered for the specified variables in the \code{Variable}
column.
}
\description{
This function is a simple helper to filter the \code{Variable} column of the
study data in long format. There are two main groups of study variables,
"original" experimental variables and "reduced" variables created in our
analyses. This functions allows to filter these two selections easily, and
optionally to filter for a specific set of variables.
}
\examples{
df_merged_long <- merge_clusters(
  df_raw     = study_data,
  df_red     = scale_reduce_vars(study_data),
  clustering = cluster_selected_vars(study_data)
) |> get_longer()

# df_merged_long contains all study and reduced variables
original <- filter_study_variables(df_merged_long, selection = "original")
reduced  <- filter_study_variables(df_merged_long, selection = "reduced")
reduced_strict  <- filter_study_variables(
 df_merged_long,
 selection = "reduced_strict"
)
custom <- filter_study_variables(
  df_merged_long,
  variables = c("VVIQ", "WCST")
)
levels(factor(original$Variable))
levels(factor(reduced$Variable))
levels(factor(reduced_strict$Variable))
levels(factor(custom$Variable))
}
