% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_glm.R
\name{fit_stan_glm}
\alias{fit_stan_glm}
\title{Fit a stan_glm model with predefined parameters}
\usage{
fit_stan_glm(
  df,
  formula,
  chains = 4,
  iter = 10000,
  refresh = 0,
  seed = 14051998,
  prior_PD = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame.}

\item{formula}{A formula specifying the model to be fitted, e.g.
\code{value ~ Group * Age}.}

\item{chains}{Number of Markov chains to run. Default is 4.}

\item{iter}{Number of iterations per chain. Default is 10000.}

\item{refresh}{Refresh rate for the progress bar. Default is 0 (no progress
bar).}

\item{seed}{Seed for random number generation to ensure reproducibility.
The default, 14051998, was used to produce the results presented in the
scientific article associated with this package.}

\item{prior_PD}{Logical. If \code{TRUE}, the model will be fitted with a prior
predictive check. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[rstanarm:stan_glm]{rstanarm::stan_glm()}}.}
}
\value{
A fitted \code{stanreg} object containing the results of the Bayesian
linear model.
}
\description{
This function is a simple wrapper around the \code{\link[rstanarm:stan_glm]{rstanarm::stan_glm()}} function
to fit a Bayesian linear model using Stan. It is mostly designed to set the
defaults we needed for our package, notably the number of iterations and the
seed for reproducibility. It provides a shorter syntax to fit many models,
allowing to reuse it in other functions (e.g., \code{\link[=get_full_model_table]{get_full_model_table()}}).
}
\examples{
df_vviq  <- study_data |> get_longer() |> dplyr::filter(Variable == "VVIQ")
vviq_fit <- fit_stan_glm(df_vviq, value ~ Group * Age)
summary(vviq_fit)
}
