% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_association_models.R
\name{format_association_table}
\alias{format_association_table}
\title{Display the association table for a specific variable in a clean format}
\usage{
format_association_table(df, variable)
}
\arguments{
\item{df}{A tibble containing the association table in a nested \code{table}
column, typically the result of \code{\link[=get_association_models]{get_association_models()}}.}

\item{variable}{A character string specifying the variable for which to
format the table. For the output of \code{\link[=get_association_models]{get_association_models()}}, this
should be one of \code{"Education"}, \code{"Field"}, or \code{"Occupation"}.}
}
\value{
A table formatted with \code{\link[knitr:kable]{knitr::kable()}}, ready to be displayed.
}
\description{
This function was built specifically to provide an easy way to display the
tables nested in the output of \code{\link[=get_association_models]{get_association_models()}} in a clean format.
It is \emph{really} not meant to be used outside of this context and it requires
specific columns, so it is not likely to work with many other data frames.
}
\examples{
df_associations <- get_association_models(study_data, group)
format_association_table(df_associations, "Education")
format_association_table(df_associations, "Field")
format_association_table(df_associations, "Occupation")
}
