% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_association_models.R
\name{get_association_models}
\alias{get_association_models}
\title{Fit and report Bayes Factors for associations with education, fields and
occupations}
\usage{
get_association_models(df, groups, type = "indepMulti")
}
\arguments{
\item{df}{A data frame containing the variables \code{education}, \code{field}, and
\code{occupation}, along with the grouping variable.}

\item{groups}{A grouping variable (e.g. \code{Group} or \code{Cluster}) without quotes.}

\item{type}{A character string specifying the type of contingency table
model to use. Default is \code{"indepMulti"}, which indicates an independent
multinomial model. See \code{?BayesFactor::contingencyTableBF} for more details
and options.}
}
\value{
A data frame summarising the Bayes Factors for associations
between the grouping variable and the variables \code{education}, \code{field}, and
\code{occupation}, with the variable name in a \code{Variable} column, the
contingency table in a \code{table} column, and the log Bayes Factor in a
\code{log_bf10} column. The variable names are capitalised for better readability.
}
\description{
This function computes the Bayes Factors for associations between a grouping
variable (e.g., \code{Group}, \code{Cluster}) and the variables \code{education},
\code{field}, and \code{occupation} in a data frame. It uses the
\code{\link[BayesFactor:contingencyTableBF]{BayesFactor::contingencyTableBF()}} function to compute the Bayes Factor
for each variable and returns a tidy data frame with the results.
}
\examples{
df_merged <- merge_clusters(
  df_raw     = study_data,
  df_red     = scale_reduce_vars(study_data),
  clustering = cluster_selected_vars(study_data)
)
get_association_models(df_merged, group)
get_association_models(df_merged, cluster)
get_association_models(df_merged, subcluster)
get_association_models(df_merged, group)$table
}
