% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bf_inclusion.R
\name{get_bf_inclusion}
\alias{get_bf_inclusion}
\title{Compute Bayes Factor for Inclusion for the groups and age covariate}
\usage{
get_bf_inclusion(df, groups, progress = FALSE)
}
\arguments{
\item{df}{A data frame in long format containing the variable to be analysed
and the required columns.}

\item{groups}{A grouping variable (e.g. \code{Group} or \code{Cluster}) without quotes.}

\item{progress}{Logical. If \code{TRUE}, the function will show a progress bar
while computing the Bayes Factor. Default is \code{FALSE}.}
}
\value{
A data frame summarising the Bayes Factor for Inclusion for the
specified grouping variable and the \code{Age} covariate, with the variable name
in a \code{Variable} column and the Bayes Factor in a \code{log_BF} column.
}
\description{
This function computes the Bayes Factor for Inclusion (BFI) for a given
grouping variable (e.g., \code{Group}, \code{Cluster}) and the \code{Age} covariate when
predicting a variable (e.g., \code{VVIQ}, \code{OSIVQ-Spatial}, etc.) from the \code{value}
column. As such, it requires a data frame with a \code{value} column, a
\code{Variable} column, the grouping variable, and the \code{Age} covariate.
}
\examples{
df_vviq  <- study_data |> get_longer() |> dplyr::filter(Variable == "VVIQ")
vviq_bfi <- get_bf_inclusion(df_vviq, Group, progress = FALSE)
print(vviq_bfi)
}
