% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrast_bf.R
\name{get_contrast_bf}
\alias{get_contrast_bf}
\title{Get contrasts between groups in a Bayesian model}
\usage{
get_contrast_bf(df, groups)
}
\arguments{
\item{df}{A data frame in long format containing the variable to be analysed
and the required columns. This is for example the output of
\code{get_longer(study_data)}.}

\item{groups}{A grouping variable (e.g. \code{Group} or \code{Cluster}) without quotes.}
}
\value{
A data frame summarising the contrasts between the levels of the
grouping variable, with the variable name in a \code{Variable} column, the
comparison in a \code{Comparison} column, the difference in a
\verb{Difference ($\\Delta$)} column, the 95\% credible interval in a
\verb{95\% CrI} column, and the log Bayes Factor in a \verb{$log(BF_\{10\})$} column.
}
\description{
This function wraps the machinery needed to compute the contrasts between the
levels of a grouping variable in our data (e.g., \code{Group}, \code{Cluster}) for a
given variable. It fits Bayesian models using the \code{\link[=fit_stan_glm]{fit_stan_glm()}} function
for the prior and posterior distributions, computes the contrasts using
\code{\link[emmeans:emmeans]{emmeans::emmeans()}} and \code{\link[=pairs]{pairs()}}, and then computes the Bayes Factor for
the contrasts using \code{\link[bayestestR:bayesfactor_parameters]{bayestestR::bf_parameters()}}. The results are then
formatted in a clean table.
}
\examples{
df_vviq <- study_data |> get_longer() |> dplyr::filter(Variable == "VVIQ")
vviq_contrast <- get_contrast_bf(df_vviq, Group)
print(vviq_contrast)
}
