% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_model_table.R
\name{get_full_model_table}
\alias{get_full_model_table}
\title{Create a table summarising the models for each variable}
\usage{
get_full_model_table(df_long, ...)
}
\arguments{
\item{df_long}{A data frame in long format containing the variables to be
analysed with a \code{Variable} column, a \code{value} column, a grouping variable
(e.g., \code{Group}, \code{Cluster}), and an \code{Age} covariate. This is for example the
output of \code{get_longer(study_data)}.}

\item{...}{A grouping variable (e.g. \code{Group} or \code{Cluster}) without quotes.}
}
\value{
A data frame summarising the models for each variable.
}
\description{
This macro-function runs various computations to model the variables in the
\code{Variable} column of a long format data that have their values in a \code{value}
column with a grouping variable (e.g., \code{Group}, \code{Cluster}) and an \code{Age}
variable. It wraps the \code{\link[=get_mean_sd]{get_mean_sd()}}, \code{\link[=get_bf_inclusion]{get_bf_inclusion()}}, and
\code{\link[=get_contrast_bf]{get_contrast_bf()}} functions to compute the mean and standard deviation
for each group, the Bayes Factor for Inclusion of the grouping variable and
the \code{Age} covariate, and the contrasts between the levels of the grouping
variable, respectively. The results are then formatted in a clean table
with one row per variable and columns for the mean, standard deviation,
Bayes Factor for Inclusion, and contrasts.
}
\examples{
df_merged <- merge_clusters(
  df_raw     = study_data,
  df_red     = scale_reduce_vars(study_data),
  clustering = cluster_selected_vars(study_data)
)
df_long_example <-
 df_merged |>
 get_longer() |>
 dplyr::filter(Variable \%in\% c("VVIQ"))

cluster_models <- get_full_model_table(df_long_example, Cluster)
print(cluster_models)
}
