% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_sd.R
\name{get_mean_sd}
\alias{get_mean_sd}
\title{Get the mean and standard deviation of a variable in a clean table}
\usage{
get_mean_sd(df, groups)
}
\arguments{
\item{df}{A data frame containing the variable to be summarised in long
format, with its name in a column and the associated values in a \code{value}
column. This is for example the output of \code{get_longer(study_data)}.}

\item{groups}{A grouping variable (e.g. \code{Group} or \code{Cluster}) without quotes.}
}
\value{
A data frame summarising the mean and standard deviation of the
variable for each group in a wide format with one column per group.
}
\description{
The point of the function is to get the mean/SD in a clean formatted string.
It does not add much on top of \code{dplyr} functions, but allows to perform the
required computations in a single line of code to use them inside other
functions (e.g., \code{\link[=get_full_model_table]{get_full_model_table()}}).
}
\examples{
df_vviq  <- study_data |> get_longer() |> dplyr::filter(Variable == "VVIQ")
vviq_mean_sd <- get_mean_sd(df_vviq, Group)
print(vviq_mean_sd)
}
