% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Add the clustering and reduced variables to the main data frame}
\usage{
merge_clusters(df_raw, df_red, clustering, names = c("B", "C", "A"))
}
\arguments{
\item{df_raw}{A data frame containing the raw data, usually \code{study_data}.}

\item{df_red}{A data frame containing the reduced variables, usually from
\code{scale_reduce_vars()}.}

\item{clustering}{A fitted Mclust object containing the clustering results,
usually from \code{cluster_selected_vars()}.}

\item{names}{A character vector of cluster names. Default is
\code{c("B", "C", "A")}. The default is based on the last iteration of the
clustering to date (which the article is based on).}
}
\value{
A data frame with the columns from the raw data, the reduced
variables, a cluster column with the cluster names and a subcluster column
with the cluster and group names.
}
\description{
Add the clustering and reduced variables to the main data frame
}
\examples{
df_merged <- merge_clusters(
  df_raw     = study_data,
  df_red     = scale_reduce_vars(study_data),
  clustering = cluster_selected_vars(study_data)
)
head(df_merged)
colnames(df_merged)
\dontshow{if (requireNamespace("dplyr")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df_merged |>
  dplyr::reframe(
    .by = cluster,
    n = dplyr::n(),
    vviq    = mean(vviq),
    object  = mean(osivq_o),
    spatial = mean(osivq_s),
    verbal  = mean(osivq_v)
   ) |>
   dplyr::arrange(cluster) |>
   print()
\dontshow{\}) # examplesIf}
}
