% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models_list}
\alias{models_list}
\title{Models for all the variables predicted by Groups, Clusters and Sub-clusters}
\format{
An object of class \code{list} of length 3.
}
\usage{
models_list
}
\description{
This list contains three data frames summarising the following:
\itemize{
\item \code{group_models}: Models for the unstandaridsed variables predicted by Groups
(aphantasia/control)
\item \code{cluster_models}: Models for the standardised variables predicted by
Clusters (A/B/C)
\item \code{subcluster_models}: Models for the standardised variables predicted by
Sub-clusters (A/B-Aphant/B-Control/C)
}
}
\details{
These tables were obtained with the \code{get_full_model_table()} function
included in the package. This saved list exists to avoid re-running the
slow computations (two Bayesian models for each variable to compute the
contrasts).
}
\keyword{datasets}
