% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters_bic.R
\name{plot_clusters_bic}
\alias{plot_clusters_bic}
\title{Plot the comparison of Gaussian Mixture Models (GMM) BIC scores}
\usage{
plot_clusters_bic(
  mclust_object,
  txt_big = 7,
  txt_mid = 6,
  txt_smol = 5,
  size = 0.2
)
}
\arguments{
\item{mclust_object}{A fitted Mclust object containing the clustering
results, usually from \code{cluster_selected_vars()}.}

\item{txt_big}{Size of the plot, x-axis and legend titles. Default is 7.}

\item{txt_mid}{Size of the legend text. Default is 6.}

\item{txt_smol}{Size of the axis text. Default is 5.}

\item{size}{Size of the points in the plot. Default is 0.2.}
}
\value{
A ggplot2 object containing the BIC plot.
}
\description{
This function is a wrapper around the \code{factoextra::fviz_mclust_bic()}
function to plot the Bayesian Information Criterion (BIC) scores of
Gaussian Mixture Models (GMM) fitted to a dataset. It is mainly designed to
facilitate the setting of various default options that fit the graphical
style of the project.
}
\examples{
study_data |> cluster_selected_vars() |> plot_clusters_bic()
}
