% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_cor_graph.R
\name{plot_score_cor_graph}
\alias{plot_score_cor_graph}
\title{Plot a correlation graph}
\usage{
plot_score_cor_graph(
  correlations,
  shape = 21,
  stroke = 0.1,
  node_size = 14,
  node_text_size = 5,
  label_text_size = 2
)
}
\arguments{
\item{correlations}{A data frame containing the correlations, typically the
output of \code{\link[=correlate_vars]{correlate_vars()}}.}

\item{shape}{Shape of the nodes in the graph. Default is 21 (circle).}

\item{stroke}{Stroke width of the nodes. Default is 0.1.}

\item{node_size}{Size of the nodes in the graph. Default is 14.}

\item{node_text_size}{Size of the text labels on the nodes. Default is 5.}

\item{label_text_size}{Size of the text labels on the edges. Default is 2.}
}
\value{
A ggplot2 object representing the correlation graph.
}
\description{
This function creates a graph visualisation of a set of correlations
computed using \code{\link[correlation:correlation]{correlation::correlation()}}, such as the result of
\code{\link[=correlate_vars]{correlate_vars()}}. It is designed to highlight nodes for specific variables
from the study (based on partial correlations), but it should run with any
set of correlations.
}
\examples{
study_data |> correlate_vars(partial = FALSE) |> plot_score_cor_graph()
}
