% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Custom ggsave wrapper set with Nature's formatting guidelines (width-locked)}
\usage{
save_ggplot(
  plot,
  path,
  ncol = 1,
  height = 90,
  show = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{plot}{The ggplot object to save.}

\item{path}{A character string with the path to save the plot.}

\item{ncol}{The number of columns for the plot. Either 1 (default) or 2.}

\item{height}{The height of the plot in mm. Default is 90 mm.}

\item{show}{Logical. Whether to return the plot visibly or not. Default
is FALSE, the plot is returned invisibly.}

\item{verbose}{Logical. Whether to print a message in the console when the
saving is done. Default is TRUE.}

\item{...}{Additional arguments passed to \code{ggsave()}.}
}
\value{
Nothing. The function saves the ggplot to the specified path.
}
\description{
See:
https://www.nature.com/documents/NRJs-guide-to-preparing-final-artwork.pdf
It's pretty strict. A one column figure is 88 mm wide and a two column figure
is 180 mm wide. Depending on the length of the figure caption, there are
different maximum heights (see the PDF). Most figures types must be in vector
format to prevent quality loss when zooming in. Ever since I found these
guidelines, I use them for all figures, even if they are not for Nature...
Because it looks nice and I like it.
}
