% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_reduce_vars.R
\name{scale_reduce_vars}
\alias{scale_reduce_vars}
\title{Reduce the number of variables to prepare for clustering}
\usage{
scale_reduce_vars(df, min = 0, max = 1)
}
\arguments{
\item{df}{A data frame containing the variables to be reduced. It should
contain the following columns:
\itemize{
\item \code{vviq}: VVIQ scores
\item \code{osivq_o}: OSIVQ Object scores
\item \code{osivq_s}: OSIVQ Spatial scores
\item \code{osivq_v}: OSIVQ Verbal scores
\item \code{psiq_vis}: Psi-Q Visual scores
\item \code{psiq_aud}: Psi-Q Auditory scores
\item \code{psiq_od}: Psi-Q Olfactory scores
\item \code{psiq_gout}: Psi-Q Gustatory scores
\item \code{psiq_tou}: Psi-Q Tactile scores
\item \code{psiq_sens}: Psi-Q Sensory scores
\item \code{psiq_feel}: Psi-Q Feelings scores
\item \code{score_raven}: Raven Matrices scores
\item \code{score_sri}: SRI scores
\item \code{span_spatial}: Spatial Span scores
\item \code{span_digit}: Digit Span scores
\item \code{wcst_accuracy}: WCST accuracy scores
\item \code{score_similarities}: Similarities scores
\item \code{score_comprehension}: Comprehension scores
}}

\item{min}{Numeric. The minimum value of the scaled range. Default is 0.}

\item{max}{Numeric. The maximum value of the scaled range. Default is 1.}
}
\value{
A data frame with reduced variables, including:
\itemize{
\item \code{visual_imagery}: Merged VVIQ, OSIVQ Object, and Psi-Q Visual scores
\item \code{auditory_imagery}: Psi-Q Auditory scores
\item \code{sensory_imagery}: Merged Psi-Q scores on other modalities
\item \code{spatial_imagery}: Merged SRI and OSIVQ Spatial scores
\item \code{verbal_strategies}: OSIVQ Verbal scores
\item \code{fluid_intelligence}: Merged Raven and Digit Span scores
\item \code{verbal_reasoning}: Similarities scores
\item \code{span_spatial}: Spatial Span scores
}
}
\description{
Reduce the number of variables to prepare for clustering
}
\examples{
df_reduced <- scale_reduce_vars(study_data)
head(df_reduced)
colnames(df_reduced)
}
