% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vars.R
\name{scale_vars}
\alias{scale_vars}
\title{Scale original quantitative variables to a defined range}
\usage{
scale_vars(df, min = 0, max = 1)
}
\arguments{
\item{df}{A data frame containing the variables to be scaled, usually
\code{study_data}.}

\item{min}{Numeric. The minimum value of the scaled range. Default is 0.}

\item{max}{Numeric. The maximum value of the scaled range. Default is 1.}
}
\value{
A data frame with the original variables scaled to the specified
range.
}
\description{
The function does not require any of the variables to be present in the
dataframe, so any dataframe can be used. The function simply modifies the
columns of interest if they
}
\examples{
df_scaled <- scale_vars(study_data, min = 0, max = 1)
head(df_scaled)
colnames(df_scaled)
}
