% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphantasiaCognitiveClustering-package.R
\docType{package}
\name{aphantasiaCognitiveClustering-package}
\alias{aphantasiaCognitiveClustering}
\alias{aphantasiaCognitiveClustering-package}
\title{aphantasiaCognitiveClustering: Data Analysis for the "Uncovering Spatial and Verbal Cognitive Profiles in Aphantasia" Study}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains all the code to reproduce the analyses reported in the article "Uncovering Spatial and Verbal Cognitive Profiles in Aphantasia Through Unsupervised Clustering".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-delem.github.io/aphantasiaCognitiveClustering/}
}

}
\author{
\strong{Maintainer}: Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

}
\keyword{internal}
