% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_cor_joint.R
\name{plot_score_cor_joint}
\alias{plot_score_cor_joint}
\title{Create a joint matrix + graph correlation figure}
\usage{
plot_score_cor_joint(correlations)
}
\arguments{
\item{correlations}{A data frame containing the correlations, typically the
output of \code{\link[=correlate_vars]{correlate_vars()}}.}
}
\value{
A ggplot2 object representing the joint plot of the correlation
matrix and the graph.
}
\description{
This function combines a correlation matrix and a graph visualisation of
correlations into a single plot using the \code{patchwork} package. While the
two underlying functions (\code{\link[=plot_score_cor_matrix]{plot_score_cor_matrix()}} and
\code{\link[=plot_score_cor_graph]{plot_score_cor_graph()}}) are fairly generic, this function is tailored for
the study data from the article. It is recommended to use it together with
the other package functions (see examples).
}
\examples{
study_data |> correlate_vars(partial = FALSE) |> plot_score_cor_joint()
}
