% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_data}
\alias{study_data}
\title{Study data retrieved from the OSF project}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 27 columns.
}
\source{
\url{https://osf.io/7vsx6/files/osfstorage}
}
\usage{
study_data
}
\description{
This dataset contains the tidied data retrieved from the OSF project. The
data has been preprocessed from the raw data files extracted from a JATOS
server beforehand. The raw data is not bundled in the package to keep it
lightweight, but the functions to extract it are available as internal
functions and a code example is included in \code{data-raw/raw_data_extraction.R}.
}
\keyword{datasets}
